// apps/web/src/lib/media/types/media.types.ts
export const MediaType = {
  AUDIO: 'audio',
  IMAGE: 'image',
  BANNER: 'banner',
  PDF: 'pdf',
  VIDEO: 'video',
  DOCUMENT: 'document',
} as const;
export type MediaType = (typeof MediaType)[keyof typeof MediaType];
// ...
export const MediaFolder = {
  AUDIO: 'audio',
  BOOKS: 'books',
  BANNERS: 'banners',
  IMAGES: 'images',
  VIDEOS: 'videos',
  AVATARS: 'avatars',
  DOCUMENTS: 'documents',
  EVENTS: 'events',
  EXPORTS: 'exports',
  TEMP: 'temp',
  PROCESSING: 'processing',
  BACKUPS: 'backups',
} as const;
export type MediaFolder = (typeof MediaFolder)[keyof typeof MediaFolder];